refactor(navigation): move gps and compass heading to service
